Movementscript as well as playerCamera
playerCamera is not finished yet, movement is closer to being done. In this the player can move forward, back, left, right with w,s,a,d and jump with space bar. The problem right now is the camera goes crazy because the object can get flipped by the thing moving